tests/requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
jsonschema==4.17.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
